Added a bunch of different options to pass a custom map function to the main BayesianOptimisation object.
Changed a couple loops in the initialisation and maximisation routines to take advantage of possible multi-threading if a suitable map function is passed as mapf. Also allowed for the possibility that in the case of a multi-threaded mapf, one might want to evaluate a couple different options during the optimisation loop at each iteration.