Introduce default printer and json observers
This commit refactors heavily how loggin is done in the package. It
removes the old style of loggin and does everything now using observers.
By default the user has access to two useful observers, json and
printer, and if the users doesn't want to deal with that the verbose
parameter takes care of the basics.