Overhaul of examples
This commit updates/fixes/improves/adds to the set of examples.

- Two tour notebooks are included with everything users need to
know about how to use the package.
- A clear yet realist example of optimizing parameters of machine
learning models is included. It is a much nicer version of the old
sklearn_example script.
- A glorious example on how to utilize this package in a concurrent
fashion was added.
- The buggy and redundant xgboost example was removed.

PS: The advanced tour notebook is not quite done yet. However I needed
to merge this branch so people could submit PRs to the updated release
branch.