Update domain_reduction.py (#454)
* Update domain_reduction.py

Adds ability to pass a dictionary as the minimum_window argument in SequentialDomainReductionTransformer. Deals with issue where TargetSpace sorts the pBounds dictionary, causing the order of the list used for minimum_window to possibly not match the order of the stored boundaries.

* Update test_seq_domain_red.py

Added test_minimum_window_dict_ordering to test whether dictionary input for SequentialDomainReductionTransformer is ordered to match pBounds.